task 1.1 0x15

+++ 0x15-file_io/1-create_file.c
@@ -24,7 +24,7 @@ int _strlen(char *str)
 int create_file(const char *filename, char *text_content)
 {
 	int fd;
-	int written, len = _strlen(text_content);
+	ssize_t written, len = _strlen(text_content);
 
 	if (!filename)
 		return (-1);
